fixed long functions

+++ src/main/java/com/kalightortaio/veterandifficulty/interfaces/IEntityState.java
@@ -10,6 +10,6 @@ public interface IEntityState {
     float getFloatState(String stateName);
     void setFloatState(String stateName, float value);
 
-    float getLongState(String stateName);
+    long getLongState(String stateName);
     void setLongState(String stateName, long value);
 }
